Subject: Added Password Match Validator
Problem: Needed a validator that checks if the password and confirm password within the auth components match.

Solution: Added a separate .ts file for the validation. Also created a reusable animation within the shared folder for labels to appear and disappear.

Notes: Will next need to check for any code cleanup opportunities and also make the auth-card component animation reuseable.